Se agrego "Ser vendedor"
