+++ WebAppPixca/Views/Productoes/Create.cshtml
@@ -2,6 +2,7 @@
 
 @{
     ViewData["Title"] = "Registro de Producto";
+    Layout = "_LayoutUser";
 }
 
 <h1>Create</h1>
