+++ WebAppPixca/Views/Shared/_LayoutUser.cshtml
@@ -1,5 +1,6 @@
 ﻿<!DOCTYPE html>
-<html lang="en">
+
+<html>
 <head>
     <meta charset="utf-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
@@ -15,7 +16,7 @@
     <header>
         <nav class="navbar navbar-expand-sm navbar-toggleable-sm navbar-light border-bottom box-shadow mb-3">
             <div class="container-fluid">
-                <img src="~/Imagenes/Mi proyecto.png" height="70" width="70"/>
+                <img src="~/Imagenes/Mi proyecto.png" height="70" width="70" />
                 <a class="navbar-brand" asp-area="" asp-controller="Home" asp-action="Index">E-Pixca Namaca</a>
                 <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target=".navbar-collapse" aria-controls="navbarSupportedContent"
                         aria-expanded="false" aria-label="Toggle navigation">
@@ -24,13 +25,26 @@
                 <div class="navbar-collapse collapse">
                     <ul class="navbar-nav d-flex ms-auto">
                         <li>
-                            <a class="nav-link text-dark" asp-area="" asp-controller="Home" asp-action="ErrorCarrito"> <i class="fa fa-shopping-cart" aria-hidden="true"></i> </a>
+                            <a class="nav-link text-dark" asp-area="" asp-controller="Usuarios" asp-action="Index"> <i class="fa fa-shopping-cart" aria-hidden="true"></i> </a>
                         </li>
-                        <li class="nav-item">
-                            <a class="nav-link text-dark" asp-area="" asp-controller="Home" asp-action="Login">Iniciar Sesión</a>
-                        </li>
-                        <li class="nav-item">
-                            <a class="nav-link text-dark" asp-area="" asp-controller="Usuarios" asp-action="Create">Registro</a>
+                        <li class="nav-item dropdown">
+                            <a class="nav-link text-dark dropdown-toggle" data-bs-toggle="dropdown">Perfil</a>
+                            <ul class="dropdown-menu dropdown-menu-end">
+                                <li class="dropdown-item">
+                                    <a class="nav-link text-dark" asp-area="" asp-controller="Usuarios" asp-action="Details">Mi perfil</a>
+                                </li>
+                                <li class="dropdown-item">
+                                    <a class="nav-link text-dark" asp-area="" asp-controller="Productoes" asp-action="Index">Mis productos</a> 
+                                </li>
+                                <hr />
+                                <li class="dropdown-item">
+                                    <a class="nav-link text-dark" asp-area="" asp-controller="Usuarios" asp-action="BeSeller">Ser vendedor</a>
+                                </li>
+                                <hr>
+                                <li class="dropdown-item">Preguntas</li>
+                                <hr />
+                                <li class="dropdown-item">@Html.ActionLink("Salir", "CerrarSesion", "Usuarios")</li>
+                            </ul>
                         </li>
                     </ul>
                 </div>
