+++ WebAppPixca/Views/Usuarios/Details.cshtml
@@ -2,6 +2,7 @@
 
 @{
     ViewData["Title"] = "Details";
+    Layout = "~/Views/Shared/_LayoutUser.cshtml";
 }
 
 <head>
@@ -15,7 +16,7 @@
     <div class="d-flex justify-content-between">
         <h1 class="mr-3">Mi perfil</h1>
         <div class="text-right">
-            <a asp-action="Index" asp-controller="Home"> <i class="fas fa-backward"></i> Regresar</a>
+            <a asp-action="HomeUser" asp-controller="Usuarios"> <i class="fas fa-backward"></i> Regresar</a>
         </div>
     </div>
     <hr />
@@ -46,6 +47,7 @@
                 <dd class="col-sm-10">
                     @Html.DisplayFor(model => model.NumeroTelefono)
                 </dd>
+                <!--
                 <dt class="col-sm-2">
                     CURP
                 </dt>
@@ -58,6 +60,7 @@
                 <dd class="col-sm-10">
                     @Html.DisplayFor(model => model.Rfc)
                 </dd>
+                -->
                 <dt class="col-sm-2">
                     Correo electrónico
                 </dt>
