+++ WebAppPixca/Views/Usuarios/Edit.cshtml
@@ -2,6 +2,7 @@
 
 @{
     ViewData["Title"] = "Edit";
+    Layout = "_LayoutUser";
 }
 
 <head>
@@ -44,7 +45,7 @@
                     <input asp-for="NumeroTelefono" class="form-control" />
                     <span asp-validation-for="NumeroTelefono" class="text-danger"></span>
                 </div>
-                <div class="form-group">
+ <!--           <div class="form-group">
                     <label asp-for="Curp" class="control-label">CURP</label>
                     <input asp-for="Curp" class="form-control" />
                     <span asp-validation-for="Curp" class="text-danger"></span>
@@ -54,6 +55,7 @@
                     <input asp-for="Rfc" class="form-control" />
                     <span asp-validation-for="Rfc" class="text-danger"></span>
                 </div>
+-->
                 <div class="form-group">
                     <label asp-for="Email" class="control-label">Correo electrónico</label>
                     <input asp-for="Email" class="form-control" type="email"/>
