+++ WebAppPixca/Views/Usuarios/HomeUser.cshtml
@@ -1,16 +1,15 @@
-﻿@*
-    For more information on enabling MVC for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
-*@
+﻿
 @{
-    ViewData["Title"] = "Home Page";
+    ViewData["Title"] = "HomeUser";
+    Layout = "_LayoutUser";
 }
 
 <head>
     <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.1.3/css/bootstrap.min.css" integrity="sha384-MCw98/SFnGE8fJT3GXwEOngsV7Zt27NXFoaoApmYm81iuXoPkFOJwJ8ERdknLPMO" crossorigin="anonymous">
     <link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.3.1/css/all.css" integrity="sha384-mzrmE5qonljUremFsqc01SB46JvROS7bZs3IO2EmfFsd15uHvIt+Y8vEf7N7fWAU" crossorigin="anonymous">
-    <link rel="stylesheet" type="text/css" href="styles.css">
 </head>
-<bodyHome class="bodyHome">
+
+<body class="bodyHome">
     <h1 class="center">Bienvenido a E-Pixca Namaca</h1>
     <p class="msj">@TempData["Mensaje"]</p>
 
@@ -18,7 +17,7 @@
         <input type="text" placeholder="Buscar..." class="inputHome">
         <buttonHome type="submit" class="buttonHome">Buscar</buttonHome>
     </form>
- 
+
     <div id="carouselExampleControls" class="carousel slide" data-bs-ride="carousel">
         <div class="carousel-inner">
             <div class="carousel-item active">
@@ -245,4 +244,6 @@
             </div>
         </div>
     </section>
-</bodyHome>
+</body>
+
+
